+++ luals-addon/factorio/factorio-plugin/arg-parser.lua
@@ -4,7 +4,7 @@ Arg parser from: https://github.com/JanSharp/LuaArgParser
 
 MIT License
 
-Copyright (c) 2021-2022 Jan Mischak
+Copyright (c) 2021-2024 Jan Mischak
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -145,12 +145,13 @@ local function parse(args, config, start_index)
 
   local function consume_args_and_convert_for_type(type_id, context)
     if not type_defs[type_id] then
-      err = "Invalid type_def id '"..type_id.."' "..context..". No type converted registered with that id."
+      error("Invalid type_def id '"..type_id.."' "..context..". No type conversion registered with that id.")
     end
     local type_args = {}
     for j = 1, type_defs[type_id].arg_count do
       if not next_arg() then
         err = "Expected "..j.." arg(s) for the type '"..type_id.."' "..context.."."
+        return
       end
       type_args[j] = current
     end
@@ -173,6 +174,7 @@ local function parse(args, config, start_index)
       value_count = value_count + 1
       values[value_count] = consume_args_and_convert_for_type(type_id, context)
     end
+    if err then return values end
     if value_count < min then
       err = "Expected "..min..(max and (max ~= min and (" to "..max) or "") or " or more")
         .." parameters of the type '"..type_id.."', got "..value_count.." "..context.."."