+++ CHANGELOG.md
@@ -10,6 +10,10 @@
   * Display the registered name of metatables, if available
 * VSCode:
   * Check LuaLS plugin files in Check Config
+* LuaLS library package:
+  * Plugin option `--ignore` to disable the plugin for certain files or folders
+  * Plugin option `--clusterio-modules` to enable handling clusterio requires
+  * Note that a bug in LuaLS 3.9.3 currently prevents passing more than one argument in some workspaces
 
 ## 1.1.43
 * Locale: